MH-236: cleaner drf serializer
